Add jp translations, move AssetContainer logic into hooks.ts, add typings
