+++ translations/ja.yml
@@ -22,9 +22,9 @@ src:
           '4035057905': 画像URL
           '4070151372': 選択
           AssetContainer:
-            '2409515782': ''
+            '2409515782': まだアセットがアップロードされていません。アップロードボタンをクリックしてファイルをアップロードしてください。
             '3057159632': 画像アップロード
-            '3715622784': ''
+            '3715622784': まだ画像がアップロードされていません。アップロードボタンをクリックして画像をアップロードしてください。
             '3863945965': ファイルアップロード
         Header:
           '248888005': アカウント設定
